requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
numpy==1.26.4
black==24.10.0
h5py==3.10.0
isort==5.13.2
joblib==1.3.2
lightning-bolts==0.7.0
matplotlib==3.8.3
opencv-python==4.9.0.80
optuna==4.0
pandas==2.2.0
pillow==10.2.0
python-javabridge==4.0.4
python-bioformats==4.1.0
pytorch-lightning==1.9.5
requests==2.32.3
scipy==1.12.0
scikit-image==0.22
scikit-learn==1.4
seaborn==0.13.2
torch==2.2.1+cu121
torchsummary==1.5.1
torchvision==0.17.1+cu121